+++ apps/maestro/src/pages/index.tsx
@@ -51,7 +51,6 @@ export default function HomePage() {
         <div className="w-full">
           <ConditionalRenderInterchainBanner />
         </div>
-
         <div
           className="flex min-h-[80dvh] w-full max-w-lg flex-col items-center justify-center"
           id="main-content"
@@ -70,7 +69,7 @@ export default function HomePage() {
               </Dialog.Body>
             </Dialog>
           )}
-          <div className="bg-base-100 grid w-full place-items-center rounded-2xl p-4">
+          <div className="grid w-full place-items-center rounded-2xl bg-base-100 p-4">
             <SearchInterchainToken onTokenFound={handleTokenFound} />
             <div className="divider w-full">OR</div>
             <InterchainTokenDeployment />