+++ packages/transaction-recovery/src/evm-execute/types.ts
@@ -1,0 +1,23 @@
+import { Environment } from "@axelarjs/core";
+
+export type EvmExecuteOptions = {
+  rpcUrl?: string;
+  privateKey?: `0x${string}`;
+};
+
+export type EvmExecuteParams = {
+  environment: Environment;
+  srcTxHash: string;
+  srcTxLogIndex?: number;
+  executeOptions?: EvmExecuteOptions;
+};
+
+export type EvmExecuteData = {
+  txHash: string;
+};
+
+export type EvmExecuteResult = {
+  success: boolean;
+  error?: string | undefined;
+  data?: EvmExecuteData | undefined;
+};